requirements.txt
fastapi[all]
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
loguru
prometheus-fastapi-instrumentator
pytest
httpx
bcrypt==4.0.1
google-genai